+++ src/app.rs
@@ -638,7 +638,7 @@ impl Handle for String {
             // I see this one when disconnecting a Bluetooth sink
             error if error == "Received error event" => {}
             // Not sure where this originates
-            error if error == "Error: Buffer allocation failed" => {}
+            error if error == "Buffer allocation failed" => {}
             _ => app.exit(Some(self)),
         }
         Ok(false) // This makes sense for now