+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - No longer exits on non-fatal PipeWire "Buffer allocation failed" errors.
+- Clear terminal after initialization.
 
 ## [0.7.0] - 2025-08-14