+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2026-01-27
+
 ### Added
 
 - Optional lazy capturing, to only monitor peak levels of on-screen nodes.